Refactor to multicommand CLI using services & handlers
